+++ src/main/java/kitchenpos/deliveryorder/infra/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.infra;
+package kitchenpos.deliveryorder.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
