+++ src/main/java/kitchenpos/eatinorder/infra/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.infra;
+package kitchenpos.eatinorder.infra;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.eatinorder.domain.OrderTable;
+import kitchenpos.eatinorder.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
