+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.application.OrderService;
-import kitchenpos.order.domain.*;
+import kitchenpos.eatinorder.application.OrderService;
+import kitchenpos.eatinorder.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
